feat: Enforce Penny Standard in DTOs and Liquidation Logic
- Converted `ProductionStateDTO.capital_stock` to `int` (pennies).
- Converted `EconomicIndicatorData.education_spending` to `Optional[int]` (pennies).
- Implemented dust-aware distribution in `LiquidationManager` to prevent monetary leakage (zero-sum guaranteed).
- Updated `test_m2_integrity` and `test_omo_system` to verify integer penny values instead of floats, aligning with the Penny Standard.
- Added `tests/test_liquidation_math.py` to verify zero-sum liquidation logic.
- Generated insight report in `communications/insights/WO-IMPL-FINANCIAL-PRECISION.md`.

+++ modules/simulation/dtos/api.py
@@ -45,7 +45,7 @@ class ProductionStateDTO:
     current_production: float
     productivity_factor: float
     production_target: float
-    capital_stock: float
+    capital_stock: int
     base_quality: float
     automation_level: float
     specialization: str
